Split seeding scripts for databases
